+++ raphtory/src/algorithms/motifs/mod.rs
@@ -1,7 +1,7 @@
-pub mod global_three_node_motifs;
+pub mod global_temporal_three_node_motifs;
+pub mod local_temporal_three_node_motifs;
 pub mod local_triangle_count;
 pub mod three_node_local_single_thread;
 pub mod three_node_motifs;
-pub mod three_node_temporal_motifs;
 pub mod triangle_count;
 pub mod triplet_count;
