+++ raphtory/src/python/packages/algorithms.rs
@@ -28,12 +28,12 @@ use crate::{
             all_local_reciprocity as all_local_reciprocity_rs,
             global_reciprocity as global_reciprocity_rs,
         },
-        motifs::global_three_node_motifs::{
+        motifs::global_temporal_three_node_motifs::{
             global_temporal_three_node_motif as global_temporal_three_node_motif_rs,
             temporal_three_node_motif_multi as global_temporal_three_node_motif_general_rs,
         },
+        motifs::local_temporal_three_node_motifs::temporal_three_node_motif as local_three_node_rs,
         motifs::local_triangle_count::local_triangle_count as local_triangle_count_rs,
-        motifs::three_node_temporal_motifs::temporal_three_node_motif as local_three_node_rs,
         pathing::{
             dijkstra::dijkstra_single_source_shortest_paths as dijkstra_single_source_shortest_paths_rs,
             single_source_shortest_path::single_source_shortest_path as single_source_shortest_path_rs,
